Creating the devkit-power branch, this brank is highly experimental a big part of xfpm and its configuration settings are broken.
In this branch polkit is a new optional dependency, this is for
getting if the user is permitted to use org.freedesktop.Devicekit.Power
sleep interfaces, also by default the power manager will try to
manipulate the lcd brightness using xrandr (doesn't work yet),
HAL remains an optional dependency of fall back for the lcd
brightness and in case of mapping failure of keyboard sleep keys.

If HAL support is compiled but HAL is not running
or stopped, xfpm should in future commits be able to handle this.