HAL has been deprecated and unmaintained for a long time now, and isn't being installed by newer distros any more, so it's time to get rid of the HAL brightness backend entirely, so drop it.
Since there are still X.org drivers which do not support brightness setting
through XRandR, add a xfpm-power-backlight-helper program which reads/writes
brightness setting in /sys/class/backlight/, and use it as a fallback if XRandR
brightness isn't available. Reading doesn't require any privileges, but for
writing the helper needs to be run as root through pkexec, so add a default
policy which allows running that without a password for the local foreground
console.

This was inspired from gnome-power-manager:

  http://git.gnome.org/browse/gnome-power-manager/commit/?id=4886023c

and the helper code comes from gpm, with slight adaptions.