Fix status icon for devices other than battery and ups (Bug 8188)
xfpm_battery_refresh_icon incorrectly adds an additional dash to
the icon prefix when the device is not a battery or a ups.

Additionally, this patch correctly shows the icon when the device
reports that it is charging.

This patch only combines the patches submitted by
Daniel Schubert and Dean Montgomery.