LXDE battery plugin support (Bug 10929)
Adds support for the battery plugin to be built and used in LXDE
as the system tray icon was removed. Thanks to
https://github.com/nandhp/multiload-nandhp as it was used as a
complete example on how to easily do this. There may be a couple
things that need to be added to make this have the same
functionality as the Xfce plugin.