Move display brightness settings to a separate tab
Brightness handling is usually done by the kernel, so the user should have the
privilege to decide whether power manager takes over the handling or not. It is
not good to automatically take over the handling, break it and then make the user
worry how to get everything going again.

This means we will need more options that will not fit on the display tab anymore,
so create an extra tab for brightness.

As an added benefit, this solves the layout problems with the sizing behaviour
of the sliders in the table.