Remove disk spin down feature (1/3)
The spin down feature does not work and would need a rewrite
to work with udisks-2. What's more, it is probably functionality
best handled by other tools or scripts, so let's remove it from
the power manager.

As a first step, this commit removes all interfacing code between
GUI and low-level udisks functions.