Add xfconf property for limiting the minimum brightness of the plugin slider
Some laptops can turn off the backlight completely by setting the brightness
to 0. Usually, this is no problem, but in case the brightness keys do not
work for whatever reason, then the user can own himself and be left in the
dark, with no easy way to regain control of the backlight.

So as a first step, let's introduce a new xfconf property to track the
minimum value the plugin slider can be set to.