Plugin: Fix crash with slider on 32bit systems (Bug 11076)
Patch provided by Simon P.

Clicking on, or dragging, the slider of the display brightness
slider of the plugin menu causes the plugin to crash. This
behaviour has been present since this feature was introduced.

I've tracked down the cause to the way the "value-changed"
signal on the scale_menu_item object is handled. The signal
is defined with a gdouble data item for the range value (but
isn't used anywhere?) while the callback function
range_value_changed_cb isn't expecting this data.

By changing to g_signal_connect_swapped for this signal in
function battery_button_show_menu & swapping the parameters
in the callback function, the range slider is now working for
me.