Don't continue to ask to start xfpm if the user cancels
If xfpm isn't running when the settings app is launched, a dialog
pops up asking if the user would like to launch xfpm. If the user
clicks cancel on that dialog, the dialog pops up again (and again).
This patch changes it so that it won't nag the user over and over.