Use translated device type for unknown devices (bug #11217)
If the device is unknown to the kernel (maybe no-name stuff or
whatever), then vendor and model will have a hex ID of 31
characters each. We do not want to show them, they are neither
useful nor human-readable, so translate and use the device
type instead of the hex IDs.

This fixes the negative side-effect that the settings window
width is too large.