Settings: Add actions for cli options
This patch adds actions to map the existing cli options. It allows
the settings dialog to be embedded as a plug and the device-id
mapping for the devices tab. Might need to rework the settings
app a little so it can active the device item rather than launch
a new window.