Use the kernel-supplied backlight type rather than a whitelist (Bug #12553)
This is a patch to replace the list of interfaces in
backlight_helper_get_best_backlight() with the same logic as here:
https://git.gnome.org/browse/gnome-settings-daemon/tree/plugins/power/gsd-backlight-linux.c#n61

i.e. grab the first interface of type "firmware" and fall back on the first
interface of type "platform" and finally "raw".

Here's the commit that replaced the list of interfaces in the GNOME source,
and the rationale:
https://git.gnome.org/browse/gnome-settings-daemon/commit/plugins/power/gsd-backlight-helper.c?id=fff09fe17eecb9314962d119a5c5b4e865f2570d

Patch by Jack Bates