panel-plugin: Set ac-adapter icon for desktop systems (Bug #13959)
Unfortunately UPower does not differentiate between desktop and laptop
systems so we decide that it's a desktop if it:
 * is not a UPS
 * does not have a battery
 * does not have a lid
In this case we show the "ac-adapter" icon for the systray and the
"ac-adapter-symbolic" icon for the Xfce Panel Plugin.

Before the "battery-missing" icon was shown, which is what UPower
reports for the display device.