Respect the notification daemon's timeouts (Fixes #57)
A longer time ago xfpm used to draw its own notifications. Setting the
timeout for all notifications is a remainder of that time.
By removing the setting of custom timeouts we comply with the standard
timeouts of the notification server.

Critical notifications are still handled as exceptions: They are set to
never expire automatically (according to the notification spec).