build: Properly distinguish X11/Wayland code paths
Many of the disabled paths are temporarily disabled, before a Wayland
implementation is added in future commits. They are therefore only
deactivated at compile time, so that you can have a test version on
Wayland by building with --disable-x11.

Everything to do with keyboard handling (brightness keys, suspend key,
etc.) is not restored on Wayland, since only the compositor is currently
able to manage this.