brightness: Rework min level management
It is now set in XfpmBrightness, so xfpm_brightness_decrease() and
xfpm_brightness_increase() can work internally as they should. This
significantly simplifies the plugin code.

This also makes the behavior of brightness keys a little more
consistent: if a custom minimum is set (this requires the systray icon
or plugin to be used), it also affects brightness keys, otherwise the
hardware minimum is used (#174).

Finally, the default custom minimum is set a little more intelligently,
depending on hardware delta and setting step.

Fixes: #174
Related: !21, !35