A review of signal handlers disconnection
In most cases, it may not make much difference, especially for the
manager who is released last and at the very end of the program. But it
may fix some crashes reported in the past, and in any case these signal
handlers should be disconnected when the object emitting the signal or
the one passed as data is released.