Always add sm-client option group
It's too early to determine the windowing system at this stage, as we
haven't yet called gtk_init(). And if we do call it, we can no longer
daemonize below, as was mentioned in a comment about
gtk_get_option_group(). So the simplest thing to do is to always add
this option group (provided you're compiling with X11 support), which
activates very little code in libxfce4ui and nothing specific to X11.

Fixes: 15180f36b7e37a09f424372a1025b002ef193ec9