Fix system-sleep-inactivity min value
If the setting was set to never before this change, this will have the
effect of setting it to 14, since this value was used as a reference for
disabling system sleep. Unfortunately, this is unavoidable and it is not
possible to implement backward compatibility, since 14 is now also a
possible value. Users will therefore have to manually set the value to 0
to disable system sleep after this change.

It's an inconvenience, but this behavior really deserved to be fixed.

Fixes: #3
(cherry picked from commit 880e9c82ef5bb64d190179d31833968d5762e145)