power: Inhibit non forced sleep requests in presentation mode too
In practice it is already the case, since this check is done in the
manager via xfpm_power_is_in_presentation_mode() before sending the
sleep request. But it's consistent with the rest of the code, i.e. with
what was done in b4f1039b. In short, presentation mode is user-set
inhibition.