settings: Rework starting xfpm if not already running
Trying to retrieve the configuration is not a valid test of xfpm's
presence on dbus, it can fail for other reasons, and this loop with
g_usleep() is not good. Better to listen to
GDBusProxy::notify::g-name-owner, with a timeout. Also, there's no point
in continuing if xfconf can't be initialized, besides the xfconf API is
used unconditionally afterwards.