systray-icon: Make setting a boolean
This was already the case in practice, since the binding in the settings
dialog stored this setting as a boolean in xfconf, and the values
defined in XfpmShowIcon were not used anywhere, except for the default
value in XfpmManager. This inconsistency ended up causing a regression
following 09a9f417, as it previously only worked by chance.

Fixes: #234