panel-plugin: Remove ungrab code when menu is shown
It's never worked (or hasn't worked for a long time) because ungrabbing
occurs too early. It can be made to work by connecting it to
GtkWidget::draw, but then it becomes almost impossible to close the menu
other than by activating an item, which would surely be seen as a big
regression by users. We can mitigate this impact while still achieving
the goal announced in comments (to enable password entry if a polkit
dialog is displayed), by ungrabbing the keyboard only. However, the menu
can no longer be closed by pressing Esc, which will also be seen as a
regression by many users (including me).

As all this is only intended to cover a rare case, and so to speak the
result of a polkit permissions misconfiguration, and as no one has
complained so far that this piece of code isn't working, we might as
well remove it, since it's causing a bug on wayland nonetheless.

Fixes: #263